revert cancellation stuff
it was difficult to come up with a test case for cancellation, implying
that implementing cancellation will have few benefits. and once I did,
implementing cancellation did not solve that test case, implying that a
real implementation will be harder than I thought. doesn't seem
cost-effective to pursue that property further for now.