fix core-lint errors
The evidence is supposed to be a core term constructing a dictionary (an
Eq# dictionary in the case of equality constraints) for the Wanted
constraint I am replacing, but I was constructing one for the Wanted
constraint I was replacing with.

Next, I should construct a dictionary for the Wanted constraint I am
replacing _in terms of_ the Wanted I am replacing it with.