do the type substitution as late as possible
instead of eagerly replacing all the type variables with their value
before performing the rewrites, only perform the substitution if we're
trying to rewrite a variable and the rewrite would have failed. this
way, we perform a lot fewer type substitutions, do ghc won't have to
re-decompose the resulting types as much.